+++ src/app/page.tsx
@@ -1,7 +1,7 @@
+'use client';
 import React from "react";
-import Homepage from "mywebsite/pages/Homepage";
 export default function Home() {
   return (
-    <Homepage></Homepage>
+    <h1>Hello</h1>
   );
 }
