+++ src/components/Navbar.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Navbar = () => {
+  return (
+    <div>Navbar</div>
+  )
+}
+
+export default Navbar
