--- src/pages/About.jsx
@@ -1,14 +1,0 @@
-'use client';
-
-import React from 'react'
-import 'mywebsite/app/globals.css'
-const About = () => {
-  return (
-    <div className='float-left'>
-        <h1>About</h1>
-          
-    </div>
-  )
-}
-
-export default About
