--- src/pages/HomePage.jsx
@@ -1,16 +1,0 @@
-'use client';
-
-import React from 'react'
-import 'mywebsite/app/globals.css'
-import ButtonRedirect from '../components/ButtonRedirect'
-const HomePage = () => {
-  return (
-    <div className='float-left'>
-        <h1>Home</h1>
-        <ButtonRedirect link={'https://github.com/MaciejPiekarczyk'} text={'Go to GitHub'} />
-        <ButtonRedirect link={'https://www.linkedin.com/in/maciej-piekarczyk-2a22bb35b/'} text={'Go to LinkedIn'} />    
-    </div>
-  )
-}
-
-export default HomePage
